Update 2-floor.py
Set return annotation to int

+++ 0x00-python_variable_annotations/2-floor.py
@@ -5,7 +5,7 @@ Define function floor
 import math
 
 
-def floor(n: float) -> float:
+def floor(n: float) -> int:
     '''
     Get the floor of a number
